requirements.txt
rasterio
shapely
geopandas
fiona
wetterdienst
pyproj
python-dateutil
fire